Added dialog to advance AST.
Added a dialog to allow advancing of AST. Checkboxes are pre-populated
based on the current civ status.

Added necessary server communication and functions to process updates.

Added necessary information on minimum requirements to advance to
Civ_Constants, an added additional static HashMap to hold the
requirement data, and code to populate the HashMap from the file.