Added timer to time phases.
Renamed the panels so that the main panel is now AstTabPanel and the
bottom portion of that panel (with the AST table) is now AstTablePanel.

Added a timer GUI to AstTabPanel to time phases (especially trading).
Beeps are set to go off at 1 min, 15s, 5s, 4s, 3s, 2s, 1s, and several
beeps when time runs out.

Unfortunately, building a timer that is run on the server may be more
trouble than it is worth trying to implement.