Created a new top-level panel to display civilization information.
Created a new top-level panel to display civilization information. Added
to the tabInformationHash to allow creation of tabs for each
civilization. 

Fixed calculation of civilization discounts in Technology.

Changed the method for the tabFactory for greater flexibility, which
also allowed me to remove the override in MegaCivFrame.

Created a method in MegaCivFrame overriding the getTabNames method to
only show available civs.

Moved civilization colors from the GUI defaults to the civilization data
file. Colors are now a static HashMap in Civilization instead of read in
as properties.

Changed loops that iterate over enums to use EnumSet.allOf() instead of
values().