Added additional sort options. (Issue 7)
Clicking on the header above the first AST column now sorts by AST
position.

Sort order for the population column now actually sorts for movement
(placing players with the Military tech at the bottom). I'm not sure
whether this is good or bad UI design (does it break with user
expectations for sorting by "population"?). Will keep it for now as it
is more economical with space, which is at a premium.