Indication of only late iron bonus; fixes to tech store. (Issue #5)
When only a single civilization is in the late iron age, the VP
displayed for that civ will have a leading asterisk, indicating it will
get the 5 VP bonus.

Made some fixes to the tech store dialog. Previously, the frame was
packed based on the first civ; if that civ had purchased many
technologies, removing the cost and thus shortening the labels, it could
cause problems when other civs are displayed. Civ labels are now a
constant size. Also changed to only three columns and removed the "next"
button.