Added an edit dialog. (Issue #2)
Created an edit dialog to allow direct editing of a civilization if
needed. Editor is reached by right-clicking on the AST Table and
choosing "Edit".

The editor creates a clone of the appropriate Civilization that can be
modified. If the edit is accepted, the new Civilization is sent to the
server to replace the old one.

Still need a method to adjust the assigned extra credits from Monument
of Written Record. This will likely require some tweaking of the
Civilization class to remember the credits for each instead of just
using a HashMap to hold the aggregate.