Changed credit model for techs that give additional credits. (Issue #2)
Changed the way additional credits from Monument and Written Record are
modeled within the Civilization object. Instead of a hashmap listing the
aggregate for each type, there is a hashmap from tech to an arraylist of
types. Only Monument and Written Record should be added to that hash.
Each type listed in the arraylist is worth 5 points. This new method
retains memory of what credits were chosen for the techs, allowing them
to be edited by the editor in a future commit.