Added credit selection to edit screen. (Issue #2)
Added combo boxes to select the extra credits for Monument and Written
Record to the CivEditDialog.

Fixed error in the Civilization constructors with regards to
initializing the extra credit field.

Created new ComboBox class for Civilization and Technology.Type and
moved the renderers inside these as private classes. Changed the
corresponding JComboBoxes to the instances of the new classes.

Adding missing breaks from server-side message handling.

Readjusted some GUI elements on the CivEditDialog to fit in 800x600.