Added ability to regress on Advance AST dialog. (Issue #4)
Added a second checkbox to the Advance AST dialog to regress the
civilization instead. This box is in the same button group as the main
checkbox, so only one can be selected (and a custom inner class
correctly allows both to be unchecked). The regress box is autochecked
if the civilization is beyond the stone age and has zero cities.

Created a new enum in Civilization to hold advancement possibilities
(Regress, None, Advance). Changed methods to take in advancement as an
argument to modify AST correctly and removed increment and decrementAST
methods.

Changed AdvanceAstMessage to carry the advancement type. Changed server
code as necessary.