Redid tech store to match civ edit screen. Fixed bug in credits.
Added combo boxes to the tech store dialog for written record and
monument, similar to those on the civ edit dialog. This eliminates the
additional dialog(s) that would appear after those techs are chosen.

Corrected a bug in the tech credit calculation in Civilization where
credits were given for written record and monument even if they weren't
owned.