Fixed issues with saving of open tabs & windows.
Fixed bug where getTabNames would actually remove non-active civs from
the tabInformationHash, causing problems if a new game was started with
different civilizations. Changed BubbaDragDropTabFrame.getTabNames to
return an ArrayList<String> copy of the set instead.

Fixed and streamline saving of open windows and tabs to the properties
file. BubbaFrame now has an abstract method createFrame(String) that
creates a frame with that name. The OpenWindows property holds the names
of the open frames. For tabbed panes, each open frame has a
Windows.<name>.OpenTabs property that holds the initial tabs.

Fixed saving of user name and controlling when it is recognized as the
default name.