Improved AST track readability. (Issue #21)
Added Apache commons-lang library. Replaced capitalizeFirst with
appropriate function call from the library.

Added AST VP to the AST step boxes when the are achieved. When a step is
achieved, the background turns the appropriate civ color (as before) and
the point value appears in the age color. Darkened the alternating age
colors for higher contrast. (Issue #21)