Implemented turn tracking.
Added methods to Game to increment and get the turn number (the field
already existed).

Added a turn counter label to the ControlsPanel to display the current
turn number.

Added field and methods to Civilization track whether the civ has
purchased techs in the current round.

Added method calls to GameServer to flip the Civilization purchased
status when a purchase message is processed.

Modified AstTabelPanel so that Buy buttons are disabled if a civ has
already purchased techs this round.