Added display for number of advances.
Added column showing number of advances to the AST table.

Added number of advances to the civ info panel.

Added tooltip to AST steps showing the requirements and whether they
have been met. Added method to Civilization to generate the string.

Added tooltip to VP column on AST table listing the sources of the
points. Added method to Civilization to generate the string.

Added tooltip to Advances column on AST table showing the number of
advances in each tier. Added method to Civilization to generate the
string.

Added visual indicator to AST table showing the first step for which the
requirements are not currently met.

Abstracted the number of VP per AST step to a static field of Game.

Changed GameServer to accept command line argument for server URL &
port.