Created mechanism to change font size based on label size.
Created a new class, LinkedLabelGroup, that will calculate the maximum
font size that fits over a number of labels, then changes the font size
for all of those labels to the new size. This allows a group of labels
that have that all have the same font size based on the dimensions of
the label.

Changed AstTablePanel to use LinkedLabelGroup to automatically change
the font size when the table is resized.