A unified timer controls all clocks on GUI.
Abstracted the stopwatch functionality out of the stopwatch panel into a
separate class. Made the new stopwatch a static field in StopwatchPanel,
so all panels share the same stopwatch. Created a StopwatchListener
interface for each stopwatch panel to register itself. StopwatchPanel
implements StopwatchListener.

The StopwatchListener interface requries four methods:
tic(int deciseconds): called whenever the stopwatch value changes
watchStarted(): called when the stopwatch is started
watchStopped(): called when the stopwatch is stopped
watchReset(): called when the stopwatch is reset

This change is also a step towards being able to have a server-based
timer.