Fixed bug where new frames did not have a client/menu.
Because the TearAwayTab was creating a new BubbaDragDropTabFrame using a
constructor, it would not create the appropriate child frame when a new
frame was created. This was fixed by removing the DnD frame constructor
and replacing it with a method that derives the new frame from an
existing frame, deriveNewFrame(). The new frame creation in TearAwayTab
was changed to use the new method. MegaCivFrame overrides the method to
copy the client as well.