Synchronized stopwatch via server.
Stopwatch actions are now synchronized via the server. This is done by
recording the connection time when the client connects and calculating
stopwatch start/stop times as an offset from that value. This should be
sufficient as long as latency is similar.

The stopwatch was moved to the client and made an instance field again.
Added a method that returns the client's stopwatch. Changed the
stopwatch panel to use the client's stopwatch and to send messages to
the server instead of manipulating the stopwatch directly.

New message classes were created to pass the stopwatch actions to and
from the server.

New methods were added to stopwatch to allow starting/stopping the timer
with an offset to account for the latency.

Currently, the client clock will not synchronize until there is a clock
set/reset. Will likely need to add a Stopwatch running on the server to
provide the information upon connecting.

This feature has not yet been fully tested. So far, this has only been
tested using multiple clients on a single computer.