Timer clock now synchronizes upon connection. (Issue #31)
Added necessary functionality to sync stopwatches on connection. This is
done by adding a field to Stopwatch that records the number of tics for
the previous event. When a client connects, it is sent three messages to
get the clock synchronized:
(1) A message to set the timer length
(2) A message to set the number of tics when the last event started
(3) If the clock is running, a message to start the clock giving the
time when the clock started.

Fixed bug where pressing reset while clock was running would not reset
the run button.