Converted Stopwatch to use util.Timer instead of Swing timer. (#31,#32)
Changed Stopwatch to use java.util.Timer instead of javax.swing.Timer
for the stopwatch thread. This gets the timer itself off of the event
dispatch thread (though the GUI updates will still be there).

Tests on the laptop show that this change allows the stopwatches to stay
in sync (on same LAN).