Added crude panel showing trade cards. (Issue #14)
Added trade card data to Civ_Constants.xml. Added parsing of that data
into Game.TRADE_CARDS.

Created TradeCardSet and TradeStack data structures to hold trade card
data. TradeCardSet holds a series of 9 TradeStack. Each TradeStack has
all of the trade cards for the associated stack number. Each
TradeCardSet holds all of the trade cards for a particular player count.

Created a rudimentary panel to display trade card data. The panel exists
on a new main panel that also displays a countdown panel. The panel as
of right now is still very basic.

Added a field to Game to track the game region being use to display the
appropriate trade cards. Modified NewGameMessage to send the additional
field, with associated changes to GameServer and NewGameDialog.

Added some improvements to when label groups are resized. One of those
is to add a change listener to the tab to resize when the tab changes.
LinkedLabelGroup also changed to not use empty string labels to
determine fontSize.

Added method calls to validate() to many updateGui to make sure the
changes are applied.