Implemented undo functionality for technology purchases. (Issue #33)
Added ability to undo technology purchases for the current round.

Changed techs field in Civilization from an ArrayList to a HashMap that
holds the technology and the round it was purchased. Altered addTech
method in Civilization to specify a round number as well as the
technology.

Added context menu item to CivRow that will undo the purchases for that
civilization in the current round. Added method to set the visibility of
this item so it can be hidden when purchases have not been made. Added
corresponding method call to AstTablePanel.updateGui() to set the
appropriate visibility.

Added undoTechPurchase method to Civilization that removes all
technologies from the specified round.

Created new class UndoPurchaseMessage to communicate an undo to the
server. Added corresponding handling to GameServer to process the new
message.