Added starting credits for small game. (Issue #34)
Added starting credits for small (5-6 player) games. This is handled by
a new HashMap in Civilization (for flexibility, right now all starting
credits are identical). A new method in Civilization,
setSmallGameCredits, which sets the credits for each tech type. There is
new data in Civ_Constants.xml that holds the credit amount, and this is
read into the static Game array SMALL_GAME_CREDITS. A new method in
Game, finalizeStart, sets the appropriate credit amount based on the
number of civs. This method is called by the server when a new game is
created after all civs have been added.

Some refactoring of fields and methods in Civilization to better
describe their function.