Added text of calamities as a tooltip.
Modified the calamity enums to contain the full calamity text. Added a
method to each that returns an html string with the calamity text.

Added tooltip to the calamity labels on the Trade tab to display the
calamity text. Ideally, this will be replaced with card images down the
line if I can get good scans.