Unused columns now hide on the Trade tab.
For single-region games, the columns not being used are now hidden on
the Trade tab. This was done by creating a new field that holds a
HashMap of the column numbers and the header labels. During updateGui,
if a panel is unused it will hide itself and the header with the same x
location. There will be redundant operations (each row will hide/show
the same header), but limits separate code.