Added tech benefits/detriments to calamity text.
Added additional text to the calamities listing modifications of the
event due to technologies, as in the rulebook. These are color-coded
with HTML tags with green for benefit and red for detriment.

Converted calamities and technologies to use new wrapHtml method for
creating word-wrapped HTML strings suitable for tooltips.

Fixed data issue where tradeable calamities were marked as non-tradeable
and vice versa.

Changed the tooltip timeout time to 5 minutes.