Added sort options to the technology purchasing dialog. (Issue #29)
Added a ComboBox to TechnologyStoreDialog that allows the user to sort
by alphabetical (previous order), base cost, or current cost (Issue
#29). This required reworking of the way the checkboxes are handled.
There is now a TechnologyCheckBox class that uses an additional field to
track the assigned Technology for the checkbox. A new method,
setCheckboxTechs(), reassigns the appropriate technologies to each
checkbox when called, tracking which techs were previously checked and
retaining that state after the sort. The previous HashMap assigning each
Technology a checkbox was converted to an ArrayList.

Added a new comparator, techCostComparator(), to Technology to allow for
sorting Technologies. If the comparator is given a Civilization, the
sort is based on the current cost, otherwise it is based on the base
cost.