Corrected bug where AST font became very small. (Issue #42)
Corrected bug where the AST font size would be set to the minimum for
the label group if one civilization satisfied the requirements for every
age (Issue #42). This was fixed by using an empty string instead of two
spaces for
empty AST boxes. LinkedLabelGroup ignores empty strings when determining
the font size. All AST labels are in the same label group, so I'm not
sure why there would be a problem if just one civ meets all
requirements, but the fix solves the problem anyway.

Also added a toString method to LinkedLabelGroup as part of the
troubleshooting.