Exclamation point now shows if reqs lost after entering age (Issue #44)
The conditional that determines whether there should be an exclamation
point indicating the requirements are not met now checks whether the
previous step was achieved as well as whether that step's requirements
were checked. This allows the mark to show properly if requirements are
lost after the age has been entered (Issue #44).