Changed AST modification dialog to show requirement status.
Updated the AST modification dialog to show the current status of each
civilization's requirements for the AST advance. Dialog now shows
whether each requirements are met using the same functionality as the
AST table mouseover. To accommodate the change, the JTextArea holding
the requirements is now set as html/text with a white background and
only the border colored by the civilization color.

Civilization.astRequirementString was modified to take an additional
argument of whether to show the age as part of the text. It was also now
correctly includes "Stone Age" (when printAgeName is true) and "No
Requirements" for Stone Age requirements.

Advancement type icons renamed to all caps to match enumeration naming
convention. Using different cases worked in the IDE, but not in
deployment.