Added shopping assistant to technology store dialog. (Issue #41)
Added a "shopping assistant" to the technology store dialog that adds
the following functionality when given a budget:
1) Adds a checkbox that will disable the purchase checkbox for any
technologies that are over budget
2) Adds a "Maximum VP" button that suggests a purchase that maximizes
the number of VP earned.

If budget is zero, these options are disabled.

These additions required the rearrangement of portions of the store
dialog. The reset button was reduced in prominence and moved towards the
center. A label displaying the total VP of the current selected
technologies was also added near the buy button.

Optimal purchase is calculated by iterating through each VP total
starting at 1. For each VP total, all combinations that result in that
total are considered and the combination that costs the least is kept as
a candidate purchase. The VP is then increased until this process
results in a minimum cost combination that does not fall within the
budget.

This commit also is (I believe) the first instance of using Java 8's
streams functionality (lambda notation). Some of the purchase cost logic
was also cleaned up some using lambda notation. Further code cleaning
will follow.