Efficiency improvements to optimal purchase procedure.
Improved the efficiency of determining optimal purchase, mainly by
restricting lists to the maximum number possible at the given VP (+
Library and Anatomy because they are special cases) and replacing some
loops with a parallelized stream. Minimum cost finding is now done
in-line on those parallel streams, resulting in less variable passing.

New procedure now completes in under a minute for most reasonable
purchasing situations (fewer than 26 VP).

Changes to the store dialog to have optimal purchases calculated off of
the EDT. While the calculation is running, the Maximal VP button is
disabled and the waiting cursor is used.

Reset button of the store dialog now also resets the budget.