Cleanup of Civilization class, mostly switching to lambda notation
Cleaned up a lot of the Civilization code, mostly to replace for loops
with streams using lambda notation.

Added a techCostComparator static class to facilitate move to lambda
notation.

Removed passAstRequirements() method as redundant with
passAstRequirements(Age age). It was only used in one place
(AstAlterationDialog), which now looks up the current age itself to use
the other method signature.

Added documentation to some of the new optimal purchase methods and
commented out some of the debugging prints.