Civ info tabs automatically open when game starts.
Client now automatically adds the trade tab and an info tab for each
civilization when a game begins, or when connecting to a server where a
game has already started. This also closes info tabs for any
civilizations not in the new game, preventing possible NPEs for those
tabs.

Added a field to BubbaDragDropTabFrame to make saving the open tabs to
file when closing optional and turned this off for MegaCiv frames. This
prevents getting NPEs when opening if the saved tabs contain civs not in
the current game (or if game is null).

Refactored the method signature for addTab because it was requiring an
redundant argument for a frame which was already covered by the object.

Added a method to BubbaGuiController that returns the first window in
the window list, which is used to auto-add the tabs when a game starts.

Fixed an issue with the classpath.