New tabs wait for user dialog to close when opening
Previously, when connecting to a game in-progress without a stored user
name, the GUI would stop to create new tabs for all the civilizations
before displaying (or displaying but not allowing entry) into the user
name dialog. Now the GUI client will wait until a real user name is
specified before adding the new tabs. This is accomplished using a
SwingWorker that checks for a user name and then sleeps for 50 ms until
one is found, at which time a new method is called to add the tabs.

Removed Trade from the starting tabs, since it will be added when the
data comes in anyway but was messing up which tab would start as visible
after tabs were added.