Removed forceUpdate argument from updateGui, documentation updates
Refactored all updateGui methods to no longer take an argument stating
whether to force an update. It was not used.

Updated documentation for AstTablePanel to produce appropriate javadoc
and clarify usage.

Choosing view from the AST Table context menu now switches to the
appropriate info tab if the current frame already has it instead of
creating a duplicate.