Moved audio from StopwatchPanel to GuiClient (Issue #39)
Moved the timer audio from StopwatchPanel into GuiClient so only one
element is trying to play sounds. This was accomplished by making the
GuiClient a StopwatchListener and adding itself as a listener to the
stopwatch. StopwatchPanel remains a stopwatch listener in order to
update the display as necessary.

Hopefully this will fix Issue #39 where sometime the sounds are not
played. No failures yet in testing.

Added a "one minute remaining" sound. Changed the filename for the final
sound.