Implemented game over logic (Issue #48), fixed AST progression bug.
Corrected a bug in AstAlterationDialog where the current step age
requirements were being used to determine whether the progress box
should be checked instead of the next step age.

Switched SortDirection.ASCENDING and SortDirection.DESCENDING, as they
were swapped incorrectly.

Updated fields and functions in Game to refer to rounds instead of
turns.

Added logic to Game to detect when the game should end (Issue #48). Each
time the round advances, if at least one civilization has reached the
late iron age, the current round (basic) or the next round (advanced) is
set as the last round. New public methods were added to determine if it
is the last round or if the game is over. The getCurrentRound method was
modified so that even though the current round is incremented past the
last round internally, it is never reported as higher than the final
round externally.

Added label to control panel that shows when it is the last turn or the
game is over. Control buttons and buy buttons are now disabled when the
game ends. The sort method of the AST Table automatically changes to VP,
Descending when the game ends.

Added a method to Civilization that gives the number of victory points
including the bonus if the civ is the only one in the late iron age.

Replaced the HashTable for civs in Game with an ArrayList. Methods were
updated as necessary for this change, using streams where appropriate.