Changed requirements not met visual to red X on AST table.
Instead of using an exclamation point, the AST Table now uses the red X
graphic to indicate the first AST step where the requirements are not
met. This is accomplished by reusing the image used by the AST
requirements string (rescaled) and using it as an Icon for the label
when appropriate.