Update to TimerMessage to provide time remaining for log (towards #36)
Added field to TimerMessage that holds the time remaining when the event
was sent. This value is used in the generation of log events.

Changed names of Stopwatch fields and methods to better describe their
usage.

Removed old constructors for Stopwatch that used an int number of
seconds instead of a Duration.