Stop clients from silently stopping updates
Add a mechanism to handle connections and send a keep-alive message to
test if connection is dead. This is done by adding a listener to the
SntpClient (which is already polling the client at a regular interval)
and sending a do-nothing message if there is an error. Hopefully, this
will eliminate instances where clients would silently stop updating.

Resolves: #50