Change how the client connection status is determined
Client is updated to determine connection state based on the connection
status of the underlying session.

When the client sends a message, the client now first checks whether it
is connected, and attempts to reconnect if it is not.

Update SNTP listener methods to include when they were triggered.

These changes will hopefully address Issue #50.

Refs: #50

Change-Id: I265af059f2283947e1c2be1d08237f2f32cb3d45