Fix save/load game functionality
Save capability was broken for multiple reasons.

The first reason was the save/load methods in GameClient were not using
the findAndRegisterModules method of ObjectMapper, so HashMaps were not
encoded properly.

The second reason is the game log includes Java 8 time objects, which
required an additional Jackson module for the parser to recognize and
encode/decode properly. Adding this new module required updating the
other Jackson modules to the same version.

Upgrading the Jackson module version led to replacing the deprecated
setVisibilityChecker method of ObjectMapper in the game message classes.

Additionally added Jackson ignore annotations to some of the static
fields in Game that do not need to be encoded in JSONs, as they never
change.

Closes: #51

Change-Id: I99b57910af5cda13c21d4d247afc97b47689e7bb