Fix add tab button in secondary frames
The add tab button was not working because the additional frame created
by dragging a tab was of type BubbaDragDropTabFrame instead of
MegaCivFrame. This caused the TabInformationHash to not be initialized
correctly. The problem is solved by a new deriveFrame method in
BubbaDragDropTabFrame that instantiates a new frame using the same
arguments, and then overriding this method in MegaCivFrame that instead
returns a MegaCivFrame.

This fix actually makes the code more straightforward, as the previous
deriveFrame method, which used to handle the drop event, is removed. The
event is instead handled in TearAwayTab, where it was generated.

Closes: #52