Improve stopwatch functionality
Streamlined the way stopwatch functions. The time remaining is now
calculated based on the last event time, rather than keeping a parallel
variable that decrements each tic. A determination of whether the
stopwatch is running is now determined by whether or not the timer is
null or not, rather than keeping a separate boolean.

The client offset from server is moved into the stopwatch itself. The
offset for the stopwatch is updated by the client whenever the sntpSync
event triggers.

The stopwatch is now responsible for generating StopwatchMessages via
the generateTimerMessage method. All building of these message outside
of Stopwatch were changed to use this method.

The StopwatchEvent was moved to stopwatch as a more logical location.
The RESET StopwatchEventType was removed as redundant and the
SET_LAST_TIC was remvoed as no longer necessary.

Stopwatch tic events are now rounded to the closest increment of
TIMER_RESOLUTION. This makes it possible to check for specific timer
thresholds to play sounds. This may have been the cause of Issue #39, as
the sounds were trying to play in rapid succession after sequential tics
would have the same number of whole seconds. Now the client checks
against the milliseconds of the tic, which are now predictable and will
not repeat.

GameServer now just rebroadcasts Stopwatch messages to all clients,
rather than regenerating its own. This means it also no longer needs to
be a StopwatchListener (but it still has to run a timer to generate
messages when a new client connects).

Closes: #39

Change-Id: I4308cd89f1c931981cd360f55ed2b1250d4b1cf6