Fix bugs with Stopwatch
Two bugs with Stopwatch were revealed upon testing on a computer with an
appreciable offset. The first is the offset was being used to correct
the time remaining, even though the client times are already corrected.
The second bug was adjusting the time before calculating the time
remaining for clock stoppages.

Change-Id: I0e94a60f2b3711897aed19570d28395bd58fa54b