Fix bug with awarding late iron age bonus
Per game creator, the Late Iron Age bonus is supposed to be awarded when
the late iron age is first reached, not a continual check. The bonus is
controlled by a new field in civilization. Game.nextRound() now checks
for this condition when the lateIronAge is first reached in expert and
awards if appropriate.

Refs: #57

Change-Id: I880aebfabebfdea507880817191c81ed0953129e