Add building scoring to game end
Support for buildings during end-game scoring is added. When the game
ends, the server sends a special message to the client who send the
final AST advance message. This message has the client GUI spawn a
dialog asking the final building disposition of each civ. When this
information is sent back to the server, new building fields in the civs
are updated, which impacts the calculated VP.

Refs: #58