Add building purchase to store dialog
Adds a checkbox to include the purchase of a special building in the
total purchase price during the technology buying phase. Building
ownership is not tracked, so this checkbox is always available.

The VP for the building are NOT reflected in the VP earned field and are
not considered in the optimal purchase calculation.

Refs: #58