Add game setting to turn off buildings
A new setting is added allowing users to choose whether to use buildings
in the game. When activated, the building purchase checkbox will appear
in the store dialog and there will be a popup at the end of the game
asking for the final building disposition. When deactivated, these will
not appear.

Refs: #58